feat: mall物品限量免费送活动高并发场景处理

+++ msa-service-mall/msa-service-mall-domain/src/main/java/top/kwseeker/msa/mall/domain/activity/model/entity/ActivityDrawEntity.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class ActivityDrawEntity {
 
     Long userId;
-    Long activityId;
+    Integer activityId;
 }
